pr-sug: rename sub-namespace to 'resiliency'
