+++ src/Arcus.Messaging.Tests.Workers.ServiceBus/MessageHandlers/CircuitBreakerAzureServiceBusMessageHandler.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 using Arcus.Messaging.Abstractions;
 using Arcus.Messaging.Abstractions.ServiceBus;
 using Arcus.Messaging.Abstractions.ServiceBus.MessageHandling;
-using Arcus.Messaging.Pumps.Abstractions.Transient;
+using Arcus.Messaging.Pumps.Abstractions.Resiliency;
 using Arcus.Messaging.Tests.Core.Messages.v1;
 using GuardNet;
